llm-agentic-vm/requirements.txt
streamlit
python-dotenv
requests
psutil
mistralai
pyyaml
openai
rich
typing-extensions
importlib-metadata
black
pylint
autopep8